NCBI --> RefSeq, many compensatory changes
