+++ conf/HymHub/Aflo.yml
@@ -4,9 +4,10 @@ Aflo:
     clade:
         - bee
         - honeybee
-    source: ncbi
-    scaffolds: afl_ref_Aflo_1.0_chrUn.fa.gz
-    annotation: ref_Aflo_1.0_top_level.gff3.gz
+    source: refseq
+    branch: invertebrate
+    accession: GCF_000184785.1
+    build: Aflo_1.0
     annotfilter:
         - 'GeneID:100869736'
         - 'GeneID:100872241'
@@ -24,7 +25,7 @@ Aflo:
         - 'GeneID:100869321'
         - 'GeneID:100868109'
     checksums:
-        gdna: 9be74682098ebb039815828cbd355ef42827b053
-        gff3: e3967d56d51a50ee58f3ab981810eb6f2d898447
+        gdna: cadd39c59a1d5b10c67c78d8c2c3e38a8aa091ae
+        gff3: d32ff75a69e5a201e26d13f0d3c1c150bf0cda01
         prot: d3817bc0b95b21980eb8aa996fc2fee9f2e0c4ac
-    description: NCBI annotation release 101
+    description: Aflo_1.0 from HGSC@Baylor
