+++ conf/HymHub/Bter.yml
@@ -4,29 +4,10 @@ Bter:
     clade:
         - bee
         - bumblebee
-    source: ncbi
-    chromosomes:
-        - bte_ref_Bter_1.0_unplaced.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B01.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B02.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B03.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B04.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B05.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B06.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B07.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B08.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B09.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B10.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B11.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B12.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B13.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B14.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B15.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B16.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B17.fa.gz
-        - bte_ref_Bter_1.0_chrLG_B18.fa.gz
-    prefix: Assembled_chromosomes/seq
-    annotation: ref_Bter_1.0_top_level.gff3.gz
+    source: refseq
+    branch: invertebrate
+    accession: GCF_000214255.1
+    build: Bter_1.0
     annotfilter:
         - 'GeneID:100648385'
         - 'GeneID:100648319'
@@ -39,7 +20,7 @@ Bter:
         - 'GeneID:100649009'
         - 'GeneID:100648960'
     checksums:
-        gdna: 729a79d46caff8f1f9c1eba829fa826c6922a4ed
-        gff3: 50f4fcadf1e032a8a86e80f8f4d5cf21549c206d
+        gdna: 43c26f7df40173119062341929c0fa07de303821
+        gff3: 73a3c220fd3f8660f07ee2f7b65b5539d7b4c63c
         prot: 826189e94fed1207a52ce387d4f132012c240549
-    description: NCBI annotation release 101
+    description: Bter_1.0 from HGSC@Baylor
