+++ conf/HymHub/Cbir.yml
@@ -2,13 +2,13 @@ Cbir:
     species: Cerapachys biroi
     common: clonal raider ant
     clade: ant
-    source: ncbi
+    source: refseq
     branch: invertebrate
     accession: GCF_000611835.1
     build: CerBir1.0
     checksums:
-        gdna: 42faab6f3ad52f4715c181423538f546fc5f3f0c
-        gff3: ae3cfb30430af10a4010bdb679eb3dd751ac2692
+        gdna: c175569906b4e7d92c011e11ec4300cf7db29dda
+        gff3: c9a3bae326030311e662d0e1519dd2782ee102aa
         prot: f92ed229aa8f371c03c4e42699fd836ca7891d6f
     description: CerBir1.0 from BGI-Shenzhen
