+++ conf/HymHub/Lhum.yml
@@ -2,12 +2,13 @@ Lhum:
     species: Linepithema humile
     common: Argentine ant
     clade: ant
-    source: ncbi
-    scaffolds: 83485_ref_Lhum_UMD_V04_chrUn.fa.gz
-    annotation: ref_Lhum_UMD_V04_top_level.gff3.gz
+    source: refseq
+    branch: invertebrate
+    accession: GCF_000217595.1
+    build: Lhum_UMD_V04
     checksums:
-        gdna: 47dec463035cad39ab01d6f808ee5aa8ebd31d03
-        gff3: 9667fab80b40d57b971391a26e5e16eefa5e5ecd
+        gdna: dc44f2c114cfb0d626084fdae054522f2d5620d8
+        gff3: 28f22718f5d03a164f5f4389f9157d204ae2666a
         prot: a3cc12377878b9c411bfcbd0a9cb9fd692d7df1e
-    description: NCBI annotation release 100
+    description: Lhum_UMD_V04 from the Ant Genomics Consortium
