+++ conf/HymHub/Pbar.yml
@@ -2,11 +2,12 @@ Pbar:
     species: Pogonomyrmex barbatus
     common: red harvester ant
     clade: ant
-    source: ncbi
-    scaffolds: 144034_ref_Pbar_UMD_V03_chrUn.fa.gz
-    annotation: ref_Pbar_UMD_V03_top_level.gff3.gz
+    source: refseq
+    branch: invertebrate
+    accession: GCF_000187915.1
+    build: Pbar_UMD_V03
     checksums:
-        gdna: 88846d721df1bf4f91700f68eb5331afaa40cb0b
-        gff3: 3ead8b1331fc766ece58a029365a69c199f672eb
+        gdna: 5deee484ae7e219b91bec975d9206378c8417f49
+        gff3: 8c39da4d76b77f3e3aaaaee30bfb0de46ba645e6
         prot: d6312f33be4a2c7d71ec08329c32f8cb4cd4ce3a
-    description: NCBI annotation release 100
+    description: Pbar_UMD_V03 from the Ant Genomics Consortium
