--- conf/test/Docc.yml
@@ -1,16 +1,0 @@
-Docc:
-    species: Draconis occidentalis
-    common: red dragon
-    source: ncbi
-    chromosomes:
-        - docc_ref_1.6_1.fa.gz
-        - docc_ref_1.6_2.fa.gz
-        - docc_ref_1.6_3.fa.gz
-        - docc_ref_1.6_4.fa.gz
-        - docc_ref_1.6_5.fa.gz
-        - docc_ref_1.6_6.fa.gz
-        - docc_ref_1.6_7.fa.gz
-        - docc_ref_1.6_8.fa.gz
-    prefix: Assembled_chromosomes/seq
-    annotation: ref_Draconis_occidentalis_1.6_top_level.gff3.gz
-    description: bogus NCBI annotation for software testing
