+++ genhub/__init__.py
@@ -18,7 +18,7 @@ import sys
 from . import conf
 from . import download
 from . import genomedb
-from . import ncbi
+from . import refseq
 from . import ncbi_flybase
 from . import beebase
 from . import crg
