+++ testdata/demo-workdir/Ador/GCF_000469605.1_Apis_dorsata_1.3_genomic.gff.gz
