+++ testdata/demo-workdir/Aech/GCF_000204515.1_Aech_3.9_genomic.gff.gz
