+++ testdata/demo-workdir/Pbar/GCF_000187915.1_Pbar_UMD_V03_genomic.gff.gz
